[concert] add category api
